+++ src/clj/cascalog/playground.clj
@@ -4,7 +4,8 @@
            [org.apache.log4j Logger WriterAppender SimpleLayout]))
 
 (defn bootstrap []
-  (use 'cascalog.api)
+  (use 'cascalog.api
+       '[jackknife.seq :only (find-first)])
   (require '(cascalog [workflow :as w]
                       [ops :as c]
                       [vars :as v])))